factor out content container
